+++ CHANGELOG.md
@@ -20,8 +20,10 @@
 |1.3.2|2017-01-07| added round arrow support |
 |1.3.3|2017-03-05| remove some unused methods |
 |1.3.4|2018-01-26| add selected text color and cell background color |
-|1.3.5|2018-06-15| Added separatorColor support. Thanks to @flyandi|
-|1.3.6|2018-06-15| Added shadow support. Thanks to @flyandi|
+|1.3.5|2018-06-15| Added separatorColor support. Thanks to @flyandi |
+|1.3.6|2018-06-15| Added shadow support. Thanks to @flyandi |
+|  | |  |
+|2.0.0|2018-07-31| supports separated configuration. |